requirements.txt
langchain
python-dotenv
streamlit==1.29.0
openai
pinecone-client
tiktoken